Renames some of the semaphore classes
