+++ jvm/src/main/java/org/pedrofelix/pc/synchronizers/UnarySemaphore.java
@@ -1,4 +1,4 @@
-package org.pedrofelix.pc.synch;
+package org.pedrofelix.pc.synchronizers;
 
 import org.pedrofelix.pc.utils.Timeouts;
 
@@ -8,31 +8,28 @@ import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
 /**
- * Semaphore with n-ary acquisition and release, and no acquisition order guarantee.
+ * Semaphore with unary acquisition and release.
  */
-public class NAnarySemaphore {
+public class UnarySemaphore {
 
     private final Lock monitor = new ReentrantLock();
     private final Condition hasUnits = monitor.newCondition();
+
     private int units;
 
-    public NAnarySemaphore(int initialUnits) {
+    public UnarySemaphore(int initialUnits) {
         units = initialUnits;
     }
 
-    public boolean acquire(int requestedUnits, long timeoutInMs)
+    public boolean acquire(long timeoutInMs)
             throws InterruptedException {
 
         monitor.lock();
         try {
 
-            if (timeoutInMs < 0) {
-                throw new IllegalArgumentException("timeoutInMs must be >=0");
-            }
-
             // fast-path (non wait-path)
-            if (units >= requestedUnits) {
-                units -= requestedUnits;
+            if (units > 0) {
+                units -= 1;
                 return true;
             }
 
@@ -44,15 +41,24 @@ public class NAnarySemaphore {
             long deadline = Timeouts.deadlineFor(timeoutInMs);
             long remaining = Timeouts.remainingUntil(deadline);
             while (true) {
+                try {
+                    hasUnits.await(remaining, TimeUnit.MILLISECONDS);
+                } catch (InterruptedException e) {
+                    // Not really needed on Java,
+                    // see https://docs.oracle.com/javase/specs/jls/se17/html/jls-17.html#jls-17.2.4
+                    if (units > 0) {
+                        hasUnits.signal();
+                    }
+                    throw e;
+                }
 
-                // No need to handle exceptions because there are not lost notifications
-                hasUnits.await(remaining, TimeUnit.MILLISECONDS);
-
-                if (units >= requestedUnits) {
-                    units -= requestedUnits;
+                // is the required condition true?
+                if (units > 0) {
+                    units -= 1;
                     return true;
                 }
 
+                // recompute remaining and check if deadline already reached
                 remaining = Timeouts.remainingUntil(deadline);
                 if (Timeouts.isTimeout(remaining)) {
                     return false;
@@ -63,12 +69,11 @@ public class NAnarySemaphore {
         }
     }
 
-    public void release(int releasedUnits) {
+    public void release() {
         monitor.lock();
         try {
-            units += releasedUnits;
-            // notice that a signalAll is required to ensure the correct thread leaves the wait-set
-            hasUnits.signalAll();
+            units += 1;
+            hasUnits.signal();
         } finally {
             monitor.unlock();
         }
